Move actor and status actions & JSP files
