+++ specimen/src/org/labkey/specimen/SpecimenModule.java
@@ -178,6 +178,18 @@ public class SpecimenModule extends SpringModule
             }
 
             @Override
+            public ActionURL getManageActorsURL(Container c)
+            {
+                return new ActionURL(SpecimenController2.ManageActorsAction.class, c);
+            }
+
+            @Override
+            public ActionURL getManageStatusesURL(Container c)
+            {
+                return new ActionURL(SpecimenController2.ManageStatusesAction.class, c);
+            }
+
+            @Override
             public ActionURL getViewRequestsURL(Container c)
             {
                 return new ActionURL(SpecimenController2.ViewRequestsAction.class, c);
